+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Adding group annotation content type
+- Add rejectDuplicates setting to workspace
 
 ## [6.2.0] - 2023-03-20
 ### Fixed